Create programming1.py
In task there was not point that it was required to perform type checking of the passed arguments, however, it would be logical for avoid an error when executing the function.
Therefore, I will offer several options for solving the task:

It was also possible to use the space (\s), but the task provided an example of the required function response, so instead of a special character, just a string with a space was used.

And more in code commend and commit